AllocationColocator: add support for arrays of non-trivial types
Summary:
In the original implementation of `AllocationColocator`, arrays of non-trivial types were omitted because initialization is a little harder and was not needed at the time.

The upcoming `ServiceInterceptor` work will need to colocate `Cpp2RequestContext` with `TypeErasedValue<T>`.

The approach to array initialize implemented here is the same as `std::generate` — a generator function.

We also need a custom deleter (encapsulated by `AllocationColocator<>::ArrayPtr`) to call the destructors of each array element. Furthermore, the creation and destruction of non-trivially typed arrays are made exception-safe, given that destructors are noexcept (`static_assert`-ed accordingly).

Lastly, `ArrayLocator<T>` now needs to carry `size` alongside `offset`. This is because we previously performed no initialization so the user was responsible for tracking the size.

Reviewed By: iahs

Differential Revision: D54831793

fbshipit-source-id: fcc899a80af5e474a368fa664ee43d7be84f99c3