Enable setting concurrencyUpperBound from existing static limit [1/n]
Summary:
`concurrencyUpperBound` setting has two main problems that have caused issues/confusion since its introduction.

First, there is a single `concurrencyUpperBound` default setting value (=65536), but there are two limiting methods (`maxRequests` and `maxQps`). This default works alright for `maxRequests` (`CONCURRENCY_LIMITS` method), since 65536 is the default `maxRequests` value in the ThriftServerConfig. However, this value makes no sense for `maxQps` (`TOKEN_BUCKET` method), which has a default ThriftServerConfig value of 0 (i.e., disabled).

The second problem, somewhat related to the first, is that there are effectively two settings parameters for controlling `maxRequests` and `maxQps`, since `concurrencyUpperBound` overwrites the value of `maxRequests` or `maxQps` at runtime. This is very confusing for users.

This change introduces a new magic value for `concurrencyUpperBound`: `kConcurrencyUpperBoundUseStaticLimit` ( = `-1`). When used, the server will use the initial `maxRequests` or `maxQps` value provided in the ThriftServerConfig as the `concurrencyUpperBound` value.

Reviewed By: praihan

Differential Revision: D54832302

fbshipit-source-id: 3fff8a0df425c66d60edc34f2d52fa785e258bd1