Refactor process_getThriftServiceMetadata method for the service
Summary:
The body of the `process_getThriftServiceMetadata()` is identical for both sync and async processors. This change moves it to a trait and use it from both processors.
It might be possible to move the code to base class `ThriftProcessorBase` for both processors, it will be done with follow up diffs.

Reviewed By: rmakheja

Differential Revision: D54856820

fbshipit-source-id: bb0b4a9bae1d59ab04a80c275132a609ad7d217b