Backout D52963494
Summary:
**Internal:** This diff reverts the changes I made in D52963494 because I no longer think the "vulnerability" that prompted the initial fix exists. I thought the issue occurred when the following operations happen within a Thrift request

1. An outgoing call is dispatched using an HTTP Thrift client, which prompts a call to setHTTPFrameworkMetadataHook
2. setHTTPFrameworkMetadataHook calls makeFrameworkMetadataHook to serialize the context into the Thrift wire format, then base64 encodes and adds the value to the write headers of to the given RpcOptions structure.
3. A component in the Thrift context is modified, eg. Artillery logs a block, Lumos adds an experiment ID etc.
4. An outgoing call is dispatched using a non-HTTP Thrift client, so the Thrift hooks serialize and place the Thrift wire format of the updated context into the outgoing request's frameworkMetadata field.
5. At this point I mistakenly assumed that we could have a stale encoded write header in the outgoing request's RpcOptions, because the Thrift context has been modified without the outgoing header reflecting the changes.

These assumptions were incorrect however, first because the headers that may contain stale encoded context strings actually belong to the RpcOptions instance of a different outgoing request. So we don't have a case where a single outgoing request sends over two different contexts in one payload, one as a header, one as the frameworkMetadata thrift field.

Also, the modification that changes the context without updating the outgoing headers happens *after* the header is computed, so the outgoing header still reflects the context that was present at the time. This is the expected behavior.

Reviewed By: praihan

Differential Revision:
D55210547

Privacy Context Container: L1180434

fbshipit-source-id: 345e73c16a9e927fa325e67c39e671bc4e7fe44e