Add iteration support to MutableStruct
Summary: Note that, as opposed to immutable thrift-python types, iterating over the mutable struct type itself (as opposed to an instance) is not implemented. This is due to the fact that we cannot cleanly define the Metaclass for a dataclass. That being said, the same outcome (i.e., listing the fields) is achievable through the (arguably more discoverable) dataclasses.fields() method, and it's unclear how useful or common iterating over the immutable type is in practice.

Reviewed By: yoney

Differential Revision: D55357965

fbshipit-source-id: 9790dde1015f3973f81e009c1b7819ff614ecb44