Fix tests around negotiating psk mode
Summary: Previously to negotiate psk_ke from SR we relied on the fact that this was the default preference order of the server. Since we've flipped the preference order on the server side (and client preference is ignored in negotiating psk mode), the only way we have to enforce psk_ke is by restricting our supported modes.

Reviewed By: knekritz, frqiu

Differential Revision: D55445608

fbshipit-source-id: 13279e793c301007657cfb4c8c729f79f82f37b7