update annotations
Summary:
Second iteration at annotating generated thrift  code.

This diff annotates *all* public client functions with data to identify the thrift generating entity, and  changes the format.

The new format to make it more general and easier to update. Using json, we can add new fields without breaking the indexer.

## Discussion

Annotating all these functions isn't very pretty. An alternative would be that the indexer uses the last comment for all non-commented functions, and annotate only the first function in the group. It would work but relies on the implicit assumption that the generated code has a specific structure. I find it slightly better to let the code generator be the source of truth and keep the indexer dumb.

Reviewed By: iahs

Differential Revision: D55445267

fbshipit-source-id: 4d22c1e95c835f16d4a34ea550d4dac27073d950