disable optimisation in RequestContextQueue constructor
Summary:
Disable optimisation for the constructor of `RequestContextQueue`. This avoids a memory corruption bug that occurs with the Xcode 15 toolchain with boost versions 1.77 - 1.79 that caused S409219.

By disabling optimisation for the constructor we ensure that the set buckets are correctly initialized as empty. This workaround can be removed once we update to Xcode 16 or boost 1.80.

Reviewed By: drodriguez

Differential Revision: D56196669

fbshipit-source-id: b690deefc59d72cb0e67b97547222f1e863876f2