InstrusiveSharedPtr
Summary:
The doc block explains the API quite well.

This is supposed to be a *mostly* drop-in replacement for `std::unique_ptr<T>` but allowing shared ownership semantics implemented via intrusive ref-counting inside `T`.

In short, it will allow us to take a safer approach to {D55506070}, which was reverted.

Reviewed By: vitaut

Differential Revision: D56043656

fbshipit-source-id: cd4cc2a6037ec4d98455dcccc1d574c79c22c44f