Make preprocess_ a collection rather than just a single function
Summary:
This diff allows to pass **multiple** preprocess function to Thrift Server.
On receiving a request the server will run these preprocess function one by one until one of them rejects the request for a reason defined in `PreprocessResult`, or all of them return `std::monostate` letting the request in.

Reviewed By: sethdelliott

Differential Revision: D55992150

fbshipit-source-id: 4f67d08775831f3121622a533389dbc5b8d27965