Require specifying name for a preprocess function when calling ThriftServer::addPreprocess
Summary: In this diff I make preprocess function to be a class (rather than just a `folly::Function`) to allow specifying name for every preprocess function. In the next diff I take advantage of this by (a) making sure there're no 2 preprocess functions with the same name in a given server (b) exposing names of all installed preprocess functions for debug purposes.

Reviewed By: sethdelliott

Differential Revision: D55992555

fbshipit-source-id: 4aca96f87d2b790dc2c2c1cf349ac882f0bbe840