Do not re-bind Protocol in serializer.pyi stub
Summary:
This pattern, of using an import alias + a non-self-referential re-bind, is one of the recommended approaches for handling re-exports as described in
https://typing.readthedocs.io/en/latest/source/stubs.html#imports

*context: Pyre needs this*

Currently Pyre uses a naive static scope extraction that cannot understand when a global is dynamically re-defined, we only handle the "last" definition.

As a result, this code (which would be a no-op at runtime in a `.py` file because `Protocol` is exactly the same value before and after) creates a cycle in Pyre's logic.

Prior to the change in D56315652 improving Pyre logic, a legacy component of Pyre with scope resolution that's inconsistent with the "normal" scope actually caused Pyre to be able to understand this code *some* of the time, but we want to delete that logic. We detected the problem when comparing trunk errors before and after D56315652.

Reviewed By: connernilsen

Differential Revision: D56403042

fbshipit-source-id: 9390f51b326a4d2117c75786c9f5ce5deabee794