Include name to make pluggable functions more debuggable
Summary:
It would be significantly easier to debug pluggable function override errors if the name of the function was included in the error message. When debugging a large number of these failures, it was very difficult to pinpoint which override was causing the problem given only a `pluggable function: override after invocation`.

To improve this experience, store the name of the pluggable function along with the impl, and print it on an override error.

Reviewed By: iahs

Differential Revision: D56523329

fbshipit-source-id: d7384f7a488504cbde806ceb45aa3a89e450b01c