Add initial MetricCollector interface
Summary:
Adds a placeholder for the new IMetricCollector interface.

Callbacks will be added to this interface to be called at specific points/milestones during the lifetime of various server-side Thrift components. This metrics-specific interface will eventually obviate the need to implement metrics instrumentation using the general purpose APIs (e.g. TServerObserver, TProcessorEventHandler, etc) which are not designed specifically for this purpose.

Reviewed By: sazonovkirill

Differential Revision: D56543238

fbshipit-source-id: 84f64c0ede4dc658d6735d6deb985314b97cd356