add cpp converter
Summary:
Right now, we have the following thrift-py3 to thrift-python migration process:
1. Make cython forward compatible with py3 converters wrappers around private API
2. Enable `auto_migrate` in binary that use the cython.
3. Codemod `.types` imports to `.thrift_types` imports to complete.

The problem is that step 3 can't be completed cleanly because the cython still depends on py3 converter.

This diff adds a thrift-python converter that can be swapped with the py3 converter in step 3. The main difference is that these functions don't force usage of `shared_ptr`. The user can choose whether to migrate to stack or `unique_ptr` (one or the other may be more efficient) or stick with shared_ptr.

Reviewed By: yoney

Differential Revision: D56589863

fbshipit-source-id: 41e07d6b61aa741ec51767c5c12e8f789f68de17