Refactor standard default value logic into separate function
Summary: Refactor the standard default value logic into a separate function `getStandardImmutableDefaultValue()` and inject it to `getDefaultValue()`. The logic (switch statement) is specific to the immutable Thrift types. For instance, the standard default value for a Thrift list is an empty `tuple`. However, for mutable types, it will be an empty `list`.

Reviewed By: aristidisp

Differential Revision: D56713436

fbshipit-source-id: c718d5790dd3cd9c9385a24b124fdda3305af5b6