Add scope parameter to requestRejected callback
Summary:
Add a `RequestRejectedScope` scope object parameter to `IMetricCollector::requestRejected()`. Having a scope object that contains the reason why the request was rejected obviates the need for indivividual callbacks per reason the request was rejected. It also sets the foundation for providing ever more rich context information about request rejected events to the `::requestRejected()` callback.

This change obviates the need for `::requestRejectedServerOverloaded()` callback, so it was removed.

Meta:

With this change, we have everything we need to add request rejected reason to the scope and implement the new per-load shedding method Fb303 counters.

Differential Revision: D56583007

fbshipit-source-id: 372e02ac3a6df86cb846c90d0cd72336c18ee074