Add API that allows to disable INFO logging on ThriftServer startup
Summary:
This diff solves the problem of extra log spew from Thrift Server in Meta's CLI tools, such as `jk`.

The problem is that CLI tools use CLF, and CLF uses Thrift Server under the hood, as an internal component. In such scenario ThriftServer's startup logs are confusing.
But we do need Thrift Server to print logs in cases when it's used as a web server (that is a typical use case).

This diff adds method `ThriftServer::disableThriftServerLogging()` that sets log level for category `thrift.lib.cpp2.server` to `FATAL`.

**NOTE**, that this only affect Folly Logging log statements (XLOG). Google Logging log statements (LOG) are **not** affected.

In the future we might consider moving all Thrift Server logs to Folly Logging. But for now this diff changes all logging statements in `ThriftServer.h` and `ThriftServer.cpp` to Folly Logging (e.g. LOG() -> XLOG).

Reviewed By: joshkehn, prakashgayasen

Differential Revision: D56554742

fbshipit-source-id: 14e5cca139df0b99dee752673d83829a5d9704d8