Merge thrift py3 and thrift python client request channel definitions into a single file
Summary: This is for the purpose of making SR backwards compatible. Cython isn't smart enough to determine that 2 extern definitions of the same cppclass are actually the same type, and SR needs to pass a request channel to whichever client it's constructing, so thrift py3 and thrift python have to share the same definition for what a client request channel is.

Reviewed By: ahilger

Differential Revision: D56998443

fbshipit-source-id: 56162fed79dee91468703c5450260079fa9537bc