CursorSerializer 6/n: add container iterator
Summary:
Allows reading containers using an input iterator without materializing the container.
This is unfortunately the strongest iterator category that doesn't require multi-pass support.
The API gets rather hairy if we want to support iteration for nested containers, so this diff only provides an API that materializes container elements. It is possible to extend this design to support nested iteration if the need arises.

Reviewed By: vitaut

Differential Revision: D56901259

fbshipit-source-id: 933cd63e96f1dc716230320efecd0cfc8593a9c7