CursorSerializer 5/n: support efficient read for string/binary types
Summary:
Supports non-copying reads for string and binary, as well as in-place copying read into std::string.
Reading into an IOBuf bumps refcounts and supports chained serialized data.
Reading into a string_view requires the serialized data to be contiguous. The allocating parser strategy added to Rocket guarantees this will be the case for RPC usecases, which we expect to be the most common usecase, but there has to be a way to turn this off for other usecases. As a result this is controlled via a template param that defaults to true for maximum usability. Failing to set it to false when required results in a runtime exception when the reader is constructed. (We don't have access to the buffer chain after the constructor, so can't defer the check until the string_view method is called.) The string_view method is disabled when Contiguous = false.

Reviewed By: vitaut

Differential Revision: D56796340

fbshipit-source-id: 9501380a708a233db17417ec4aa05665d403304f