Make naming of CPUConcurrencyController::Method enum values more accurate
Summary:
We want to restore the original meaning of ThriftServer::setMaxRequests, where ThriftServer::setMaxRequests means the total number of requests in the system, including queued and executing requests. That is necessary to give customers the capability to protect their servers from OOMs back. We lost that capability by accident while working on ResourcePools.

In preparation for that work I want to change the naming of CPUConcurrencyController::Mode options to match their meaning more accurately.

  Current                     New
  CONCURRENCY_LIMITS          MAX_REQUESTS
  TOKEN_BUCKET                MAX_QPS

This is necessary because we will need to add ThriftServer::setConcurrencyLimit() to allow CPU-CC to control that property in the future.

This is a non-trivial change because it requires changing both fbcode and configerator definitions.

### Plan
- Fbcode change that simply renames options of `CPUConcurrencyController::Mode`. Definitions in configerator stay unchanged. (THIS DIFF)
- Configerator change:
  - Add new fields `CPUController.cpuControllerMode` and `CPUController.cpuControllerMethod` to eventually replace `CPUController.mode` and `CPUController.dryRun`.
  - Add logic that makes it so that new fields always match old fields. **D56985934**
  - Adjust C++ code to always read from new fields
  - Replace old fields with new fields in Configerator configs.
  - "Erase" old fields from `internal_overload_config.thrift`

Reviewed By: sethdelliott

Differential Revision: D56985373

fbshipit-source-id: 88371c0154492964c287cb3b3ce01895b202789b