resize list for trivial types only
Summary:
Address comments in D56663331, resize trivially_constructible list might not always be optimum (e.g. for list<shared_ptr>).

Quote from ot
```
imagine elem_type is std::shared_ptr, then it is trivially constructible,
but in order to assign to it you first need to read it back to check if it
is non-empty and needs to be destroyed, which is much more expensive
than constructing in-place.
```

Reviewed By: thedavekwon

Differential Revision: D57161529

fbshipit-source-id: 6a95fe6af7ae112034216b5fca68be9953ff4cef