Ensure wangle::TransportInfo securityType is correct for StopTLS
Summary:
`wangle::TransportInfo` `securityType` should be a copy of the handshake layer's
security protocol by the time the application `onNewConnection` method is invoke.

This is the case for the normal OpenSSL handshake and FizzAcceptorHandshakeHelper.

Thrift adds another step for StopTLS handshakes, and StopTLS is implemented as a
separate transport with a unique security type, so this diff ensures that
`securityType` is still accurate.

Reviewed By: stuclar

Differential Revision: D57394275

fbshipit-source-id: f8051de3598a8d8e8f84bce5ea58c6e6c66eeb0a